+++ src/root/views/CountryProfile/IndicatorOutput.js
