+++ src/root/views/CountryProfile/InformIndicators.js
