+++ src/root/views/CountryProfile/NSIndicators.js
